+++ Sources/Sentry/SentryNetworkTracker.m
@@ -58,6 +58,8 @@ SentryNetworkTracker ()
             return;
         }
     }
+    
+    if (![self isTaskSupported:sessionTask]) return;
 
     // We need to check if this request was created by SentryHTTPInterceptor so we don't end up with
     // two spans for the same request.
@@ -73,7 +75,7 @@ SentryNetworkTracker ()
 
     NSURL *url = [[sessionTask currentRequest] URL];
 
-    if (url == nil || ![self isTaskSupported:sessionTask])
+    if (url == nil)
         return;
 
     // Don't measure requests to Sentry's backend